Extract object_template.h and .cc
