+++ v8go.h
@@ -104,12 +104,6 @@ extern int TemplateSetAnyTemplate(TemplatePtr ptr,
                                   TemplatePtr obj_ptr,
                                   int attributes);
 
-extern TemplatePtr NewObjectTemplate(IsolatePtr iso_ptr);
-extern RtnValue ObjectTemplateNewInstance(TemplatePtr ptr, ContextPtr ctx_ptr);
-extern void ObjectTemplateSetInternalFieldCount(TemplatePtr ptr,
-                                                int field_count);
-extern int ObjectTemplateInternalFieldCount(TemplatePtr ptr);
-
 extern TemplatePtr NewFunctionTemplate(IsolatePtr iso_ptr, int callback_ref);
 extern RtnValue FunctionTemplateGetFunction(TemplatePtr ptr,
                                             ContextPtr ctx_ptr);
